requirements.txt
# ==== Core Machine Learning & Deep Learning ====
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

# ==== Data Handling ====
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0

# ==== Image Processing ====
pillow==10.2.0
opencv-python==4.9.0.80
albumentations==1.4.4
imageio==2.34.0

# ==== Visualization & Analysis ====
matplotlib==3.8.3
seaborn==0.13.2
tqdm==4.66.2

# ==== Utility & Environment ====
jsonschema==4.21.1
jupyter==1.0.0
notebook==7.1.0

# ==== Metrics ====
scikit-image==0.22.0
plotly==5.19.0
